refactor : 폴더 및 로직 수정
